requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
pyjwt==2.8.0
httpx==0.25.2

alembic>=1.7.0
python-multipart>=0.0.5